+++ packages/wiz-ui-react/src/components/base/inputs/search-input/stories/dummy-data.ts
